Move ADBaseShapeProvider from adcore io to logic
